Move CStringArray into SmingCore/Data
